auth/login/requirements.txt
requests == 2.31.0
bcrypt == 4.1.2
PyJWT == 2.8.0